Initial Commit
This is the final source for Sammy The Shark's Node.JS Particle Server

This uses JohnnyFive, and custom SparkIO to communicate with a local
cloud via WiFi. The ParticleCore is flashed with Voodoo firmware.

On the client side, WebSockets is used to communicate with WS for Node.
HTML5 deviceorientation API and TouchEvents is used for the controls.

Some extra work was done to store lap times but was abandoned (originally
for the Digital Ocean Offline First Hackathon)